repackage to package by component with separate data package
